requirements.txt
discord.py
dnspython==1.16.0
PyNaCl
async-timeout==3.0.1
youtube-dl==2020.09.20
pandas
pyfiglet
sympy
giphy_client
